ml/backend/app-python/requirements.txt
annotated-types==0.7.0
fastapi==0.115.0
fastapi-cli==0.0.5
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
starlette==0.38.5
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
sqlalchemy==2.0.35
werkzeug==3.0.1
psycopg==3.2.2
PyMuPDF
python-docx
langchain
langchain-community
python-multipart==0.0.12
boto3==1.35.34